Merge upstream changes 644c8a88764746ed029d858c0ddaa5ec33975ff7
msm: mdp3: clear continuous splash flag after splash completion

Commits 1075e52 and
16dd03e introduced changes
which prevents backlight updation when continuous splash is in
progress. The continuous splash flag (cont_splash_enabled) was
not getting cleared to zero even after the continuous splash was
over. Due to this, the backlight updation was failing both in
normal and recovery mode. Fix this by clearing the
cont_splash_enabled flag once the splash is over.

Change-Id: I6f4192a437731ff70aa600d86f47f6069bac763e
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>